requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
astroid==3.0.3
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
decorator==5.1.1
dicttoxml==1.7.16
dill==0.3.8
docutils==0.20.1
exceptiongroup==1.2.0
flask==3.0.1
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
frozenlist==1.4.1
future==0.18.3
geocoder==1.38.1
geoip2==4.8.0
google-analytics-data==0.18.6
google-api-core==2.17.1
google-auth==2.28.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
ip2geotools==0.1.6
IP2Location==8.10.2
isort==5.13.2
itsdangerous==2.1.2
jaraco.classes==3.3.1
Jinja2==3.1.3
keyring==24.3.0
lazy-object-proxy==1.10.0
lxml==5.1.0
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
maxminddb==2.5.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
nh3==0.2.15
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pip-review==1.3.0
pkginfo==1.9.6
platformdirs==4.2.0
plotly==5.18.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pygments==2.17.2
pylint==3.0.3
pyparsing==3.1.1
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
pywin32-ctypes==0.2.2
ratelim==0.1.6
readme-renderer==42.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
rsa==4.9
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
tenacity==8.2.3
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
twine==5.0.0
typed-ast==1.5.5
typing==3.7.4.3
typing-extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.1.0
waitress==2.1.2
webencodings==0.5.1
werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
